Fixes audit log creation for 2fa enforcement

+++ src/sentry/tasks/auth.py
@@ -9,10 +9,12 @@ from django.urls import reverse
 from sentry import audit_log, features, options
 from sentry.auth import manager
 from sentry.auth.exceptions import ProviderNotRegistered
-from sentry.models import ApiKey, AuditLogEntry, Organization, OrganizationMember, User, UserEmail
+from sentry.models import Organization, OrganizationMember, User, UserEmail
 from sentry.services.hybrid_cloud.organization import organization_service
+from sentry.services.hybrid_cloud.user import RpcUser
 from sentry.services.hybrid_cloud.user.service import user_service
 from sentry.tasks.base import instrumented_task, retry
+from sentry.utils.audit import create_audit_entry_from_user
 from sentry.utils.email import MessageBuilder
 from sentry.utils.http import absolute_uri
 
@@ -81,12 +83,18 @@ class OrganizationComplianceTask(abc.ABC):
     def remove_non_compliant_members(
         self, org_id: int, actor_id: int | None, actor_key_id: int | None, ip_address: str | None
     ):
-        actor = User.objects.get(id=actor_id) if actor_id else None
-        actor_key = ApiKey.objects.get(id=actor_key_id) if actor_key_id else None
+        actor = user_service.get_user(user_id=actor_id) if actor_id else None
 
-        def remove_member(member):
-            user = user_service.get_user(user_id=member.user_id)
-            logging_data = {"organization_id": org_id, "user_id": user.id, "member_id": member.id}
+        def remove_member(org_member: OrganizationMember):
+            user = user_service.get_user(user_id=org_member.user_id)
+            if user is None:
+                return
+
+            logging_data = {
+                "organization_id": org_id,
+                "user_id": user.id,
+                "member_id": org_member.id,
+            }
 
             removed_member = organization_service.remove_user(
                 organization_id=org_id, user_id=user.id
@@ -100,16 +108,17 @@ class OrganizationComplianceTask(abc.ABC):
                 logger.info(
                     f"{self.log_label} noncompliant user removed from org", extra=logging_data
                 )
-                AuditLogEntry.objects.create(
-                    actor=actor,
-                    actor_key=actor_key,
+                create_audit_entry_from_user(
+                    user=actor,
+                    actor_key_id=actor_key_id,
                     ip_address=ip_address,
                     event=audit_log.get_event_id("MEMBER_PENDING"),
-                    data=member.get_audit_log_data(),
+                    data=org_member.get_audit_log_data(),
                     organization_id=org_id,
                     target_object=org_id,
                     target_user_id=user.id,
                 )
+
                 org = Organization.objects.get_from_cache(id=org_id)
                 self.call_to_action(org, user, member)
 
@@ -129,7 +138,7 @@ class TwoFactorComplianceTask(OrganizationComplianceTask):
             return user.has_2fa()
         return False
 
-    def call_to_action(self, org: Organization, user: User, member: OrganizationMember):
+    def call_to_action(self, org: Organization, user: RpcUser, member: OrganizationMember):
         # send invite to setup 2fa
         email_context = {"url": member.get_invite_link(), "organization": org}
         subject = "{} {} Mandatory: Enable Two-Factor Authentication".format(
@@ -167,7 +176,7 @@ class VerifiedEmailComplianceTask(OrganizationComplianceTask):
             return UserEmail.objects.get_primary_email(user).is_verified
         return False
 
-    def call_to_action(self, org: Organization, user: User, member: OrganizationMember):
+    def call_to_action(self, org: Organization, user: RpcUser, member: OrganizationMember):
         import django.contrib.auth.models
 
         if isinstance(user, django.contrib.auth.models.User):
